// fe/src/lib/api/generated/model/accountCreateRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The version-one backend API for COMPASS.
 * OpenAPI spec version: 1.0.0
 */
import type { RoleCode } from './roleCode';

export interface AccountCreateRequest {
  email: string;
  first_name: string;
  institutional_id: string;
  is_active?: boolean;
  last_name: string;
  middle_name?: string;
  role: RoleCode;
  suffix?: string;
}